SF extract constants out of internal
