+++ modules/apps/oauth2-provider/oauth2-provider-web/src/main/java/com/liferay/oauth2/provider/web/constants/OAuth2ProviderPortletKeys.java
@@ -12,25 +12,22 @@
  * details.
  */
 
-package com.liferay.oauth2.provider.web.internal.constants;
+package com.liferay.oauth2.provider.web.constants;
 
 /**
  * @author Stian Sigvartsen
  */
-public class OAuth2ProviderWebKeys {
+public class OAuth2ProviderPortletKeys {
 
-	public static final String ASSIGN_SCOPES_MODEL = "ASSIGN_SCOPES_MODEL";
+	public static final String OAUTH2_ADMIN =
+		"com_liferay_oauth2_provider_web_internal_portlet_OAuth2AdminPortlet";
 
-	public static final String AUTHORIZATION_MODEL = "AUTHORIZATION_MODEL";
+	public static final String OAUTH2_AUTHORIZE =
+		"com_liferay_oauth2_provider_web_internal_portlet_" +
+			"OAuth2AuthorizePortlet";
 
-	public static final String OAUTH2_ADMIN_PORTLET_DISPLAY_CONTEXT =
-		"OAUTH2_ADMIN_PORTLET_DISPLAY_CONTEXT";
-
-	public static final String OAUTH2_AUTHORIZE_PORTLET_DISPLAY_CONTEXT =
-		"OAUTH2_AUTHORIZE_PORTLET_DISPLAY_CONTEXT";
-
-	public static final String
-		OAUTH2_CONNECTED_APPLICATIONS_PORTLET_DISPLAY_CONTEXT =
-			"OAUTH2_CONNECTED_APPLICATIONS_PORTLET_DISPLAY_CONTEXT";
+	public static final String OAUTH2_CONNECTED_APPLICATIONS =
+		"com_liferay_oauth2_provider_web_internal_portlet_" +
+			"OAuth2ConnectedApplicationsPortlet";
 
 }
