+++ modules/apps/oauth2-provider/oauth2-provider-web/src/main/resources/META-INF/resources/admin/init.jsp
@@ -36,10 +36,10 @@ page import="com.liferay.oauth2.provider.scope.liferay.LiferayOAuth2Scope" %><%@
 page import="com.liferay.oauth2.provider.service.OAuth2ApplicationScopeAliasesLocalService" %><%@ 
 page import="com.liferay.oauth2.provider.service.OAuth2ApplicationScopeAliasesLocalServiceUtil" %><%@
 page import="com.liferay.oauth2.provider.service.OAuth2AuthorizationServiceUtil" %><%@ 
-page import="com.liferay.oauth2.provider.web.internal.constants.OAuth2ProviderWebKeys" %><%@
+page import="com.liferay.oauth2.provider.web.constants.OAuth2ProviderWebKeys" %><%@
 page import="com.liferay.oauth2.provider.web.internal.display.context.AssignScopesModel"%><%@
 page import="com.liferay.oauth2.provider.web.internal.display.context.AuthorizationModel"%><%@
-page import="com.liferay.oauth2.provider.web.internal.constants.ClientProfile"%><%@
+page import="com.liferay.oauth2.provider.web.constants.ClientProfile"%><%@
 page import="com.liferay.oauth2.provider.web.internal.display.context.OAuth2AdminPortletDisplayContext" %><%@
 page import="com.liferay.oauth2.provider.web.internal.display.context.OAuth2ApplicationsManagementToolbarDisplayContext" %><%@
 page import="com.liferay.oauth2.provider.web.internal.display.context.OAuth2AuthorizationsManagementToolbarDisplayContext" %><%@
