+++ modules/apps/oauth2-provider/oauth2-provider-web/src/main/resources/META-INF/resources/authorize/init.jsp
@@ -16,7 +16,7 @@
 <%@ include file="/init.jsp" %>
 
 <%@ page import="com.liferay.oauth2.provider.service.OAuth2ApplicationLocalServiceUtil" %><%@
-page import="com.liferay.oauth2.provider.web.internal.constants.OAuth2ProviderWebKeys"%><%@
+page import="com.liferay.oauth2.provider.web.constants.OAuth2ProviderWebKeys"%><%@
 page import="com.liferay.oauth2.provider.web.internal.display.context.AuthorizationModel"%><%@
 page import="com.liferay.oauth2.provider.web.internal.display.context.OAuth2AuthorizePortletDisplayContext" %><%@
 page import="com.liferay.petra.string.StringPool"%><%@
