+++ modules/apps/oauth2-provider/oauth2-provider-web/src/main/resources/META-INF/resources/connected_applications/init.jsp
@@ -27,7 +27,7 @@ page import="com.liferay.portal.kernel.util.Validator" %><%@
 page import="com.liferay.portal.kernel.util.StringUtil" %><%@
 page import="com.liferay.oauth2.provider.web.internal.display.context.OAuth2ConnectedApplicationsPortletDisplayContext" %><%@
 page import="com.liferay.oauth2.provider.web.internal.display.context.AuthorizationModel" %><%@
-page import="com.liferay.oauth2.provider.web.internal.constants.OAuth2ProviderWebKeys" %>
+page import="com.liferay.oauth2.provider.web.constants.OAuth2ProviderWebKeys" %>
 
 <%
 OAuth2ConnectedApplicationsPortletDisplayContext oAuth2ConnectedApplicationsPortletDisplayContext = (OAuth2ConnectedApplicationsPortletDisplayContext)request.getAttribute(OAuth2ProviderWebKeys.OAUTH2_CONNECTED_APPLICATIONS_PORTLET_DISPLAY_CONTEXT);
